BST-244 Ensure AVL root node is correctly managed by AvlTree
When the AVL tree rebalances on insertion, the root node of the tree may
change, and this needs to be managed by the AvlTree class. Along with
ensuring correct root node reassignment:

- AvlTree specs updated.

- an AvlTree spec file with examples was deleted after specs moved
to AvlTree specs.

- AvlTree class massively shortened, much code deleted.
